requirements.txt
 aiohttp==3.9.5
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.6.2
Bottleneck==1.3.7
Brotli==1.0.9
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.4
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
clickhouse-connect==0.7.15
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.5
dataclasses-json==0.6.5
Deprecated==1.2.13
distro==1.9.0
duckdb==1.0.0
exceptiongroup==1.2.0
fastapi==0.103.0
filelock==3.13.1
flatbuffers==2.0
frozenlist==1.4.0
fsspec==2024.3.1
gmpy2==2.1.2
google-auth==2.29.0
googleapis-common-protos==1.56.4
greenlet==3.0.1
grpcio==1.62.2
h11==0.14.0
hnswlib==0.7.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
huggingface_hub==0.23.1
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.1
importlib-resources==6.1.1
jsonpatch==1.33
jsonpointer==2.1
kubernetes==29.0.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.11
langchain-text-splitters==0.2.2
langsmith==0.1.84
lz4==4.3.2
markdown-it-py==2.2.0
marshmallow==3.19.0
mdurl==0.1.0
mmh3==4.1.0
monotonic==1.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.35.10
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.9.15
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pip==24.0
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydantic-settings==2.3.4
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==24.0.0
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==1.3.0
rich==13.3.5
rsa==4.7.2
setuptools==69.5.1
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.27.0
sympy==1.12
tenacity==8.5.0
tokenizers==0.15.1
tomli==2.0.1
tqdm==4.66.4
typer==0.9.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.19
uvicorn==0.20.0
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
wrapt==1.14.1
yarl==1.9.3
zipp==3.17.0
zstandard==0.22.0
